requirements.txt
alsa-midi
RPi.GPIO